Refactor to prevent 'was actually of type jdk.proxy2.$Proxy221' issue.
